refactor: moved get open issues to just home bc it does not need to run on every page. should improve performance. status banner will not be in preview link bc server action cant be static
