+++ app/layout.tsx
@@ -1,18 +1,9 @@
 import { Inter, Source_Sans_3, Source_Serif_4 } from "next/font/google";
 import "@/app/globals.css";
 import { ReactNode } from "react";
-import { getOpenIssues } from "@/lib/get-open-issues";
-import { unstable_cache } from "next/cache";
-import LayoutWithStatusBanner from "@/components/LayoutWithStatusBanner";
 import BrownBanner from "@/components/BrownBanner";
 import Navbar from "@/components/Navbar";
 import Footer from "@/components/Footer";
-
-const getCachedOpenIssues = unstable_cache(
-  getOpenIssues,
-  ["open-issues"],
-  { revalidate: 60 }
-);
 
 const inter = Inter({ subsets: ["latin"] });
 const sourceSans = Source_Sans_3({
@@ -41,8 +32,6 @@ export default async function RootLayoutWrapper({
 }: {
   children: ReactNode;
 }) {
-  const issues = await getCachedOpenIssues();
-
   return (
     <html
       lang="en"
@@ -55,7 +44,6 @@ export default async function RootLayoutWrapper({
         className={`${inter.className} m-0 p-0 overflow-x-hidden bg-white`}
       >
         <div>
-          <LayoutWithStatusBanner issues={issues} />
           <BrownBanner />
           <Navbar />
           <div className="flex-grow w-full">{children}</div>
